fix: consolidate keyboard navigation hook
